refactor: 呼び出し元の import を utils shim から domain 直接へ移行
- hooks/useOmikujiLogic: drawOmikuji / canDrawToday を domain から import、
  lastDate === getTodayString() の比較を canDrawToday(lastDate) に置き換え
  (未使用だった domain API の利用を確立)
- components/patterns/ResultPattern: buildShareText / getFortuneText を
  domain から import
- components/patterns/HistoryListPattern: getFortuneText を domain から
  import (HistoryEntry 型は副作用を含む utils/HistoryStorage のまま)

utils/ の @deprecated shim は互換目的で当面残置。全呼び出し元が移行した
段階で削除フォローアップを計画中 (別 PR)。

テスト結果: lint / tsc / test いずれも green
(43 suites / 286 tests passed, 挙動変更なし)
